refactor(maps): collapse pipeline into make_map.py + tdmap.py + regions.py
Issue #132. The host-side map generator was 8 Python files
(pmtiles_to_tdmap.py + config + process + archive + land_mask + labels +
render, plus planetiler.sh) -- a tour-of-five-files for anyone wanting
to make their own region. Collapse to one user-facing CLI plus one
format module:

  * make_map.py -- single CLI:
      make_map.py <region>                            (preset)
      make_map.py custom <pmtiles> --bounds --zoom    (one-off)
      make_map.py inspect <tdmap>                     (debug)
  * tdmap.py    -- merged: TDMAPWriter/Reader, MVT renderer, label
      extractor, land mask, 3-bit packing, inspect/verify CLI.
  * regions.py  -- region presets dataclass with the planetiler
      filename convention baked in.

Source PMTiles preparation stays in planetiler.sh, documented as the
one-time setup step (Protomaps daily builds aren't usable as a
per-region URL -- range-extract over HTTP would mean thousands of
requests against their CDN, which their docs explicitly discourage).

Loud failures replace silent empty maps:
  * source PMTiles missing  -> prints exact planetiler.sh invocation
  * zoom outside source     -> "requested zoom A..B is outside M..N"
  * bounds produce 0 tiles  -> "bounds produce zero tiles..."

Checkpoint behaviour is now invisible by default: the .checkpoint
sidecar is dropped during the run, validated and re-used on rerun if
config + source still match, and removed on success.

TDMAP v6 format unchanged; reader unchanged; on-device renderer
untouched. Net diff: -2,859 lines, 7 files deleted, 3 added.
